Lots of updates
Wave system functional for practically unlimited waves (alternating between incrementing enemy count and enemy health until reaching max health, then switch to incrementing speed instead of health).
Added a display to show what wave is starting.
Added a delay when the wave number is being displayed also to give the player a respite between waves.
Tweaks to code to stop null reference exceptions from showing up when changing Enemy Sprites.
Name input added and name requirement to start game added to code.
Made all Managers inherit from GameManager and removed the need to have it be present as a GameObject.
Made Awake() a virtual method to ensure proper resets when a new scene is loaded (so far just resetting TimeScale).
Overriding that Awake() for functionality specific to the respective scenes.
Added crosshairs and an indicator of whether the portal is available that follow the mouse.
Hid the cursor and confined it to the game window.
Fixed Stun() so that a new Stun() call before the first one resolves renews the timer instead of not doing that (it actually works by counting the number of active Stun() calls and only resolving them if there's only one active, otherwise it does nothing and decrements the number of active Stun() calls).
Made a GameOver screen and a method to load that and hide the Pause screen as well as updating the Pause method to do vice versa.